requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=./requirements.txt requirements.in
#
attrs==19.1.0
python-dotenv==0.10.3
